Move Mix task preflight checks to a module
